styled tags in select component
